// test/integration/test/ports/_fixtures/composites-object-create/generated/contract.d.ts
export type CodecTypes = MongoCodecTypes;
// ...
export type FieldOutputTypes = {
  readonly __unbound__: {
    readonly CommentOptional: {
      readonly _id: CodecTypes['mongo/objectId@1']['output'];
      readonly country: CodecTypes['mongo/string@1']['output'] | null;
      readonly content: CommentContentOutput | null;
    };
    readonly CommentRequired: {
      readonly _id: CodecTypes['mongo/objectId@1']['output'];
      readonly country: CodecTypes['mongo/string@1']['output'] | null;
      readonly content: CommentContentOutput;
    };
  };
};
export type FieldInputTypes = {
  readonly __unbound__: {
    readonly CommentOptional: {
      readonly _id: CodecTypes['mongo/objectId@1']['input'];
      readonly country: CodecTypes['mongo/string@1']['input'] | null;
      readonly content: CommentContentInput | null;
    };
    readonly CommentRequired: {
      readonly _id: CodecTypes['mongo/objectId@1']['input'];
      readonly country: CodecTypes['mongo/string@1']['input'] | null;
      readonly content: CommentContentInput;
    };
  };
};
export type TypeMaps = MongoTypeMaps<CodecTypes, FieldOutputTypes, FieldInputTypes>;
// ...
export type Contract = MongoContractWithTypeMaps<ContractBase, TypeMaps>;